feat: Add tailwind, react-router-dom
I decided to use tailwind in this project to
know it better. I was thinking about write the
entire css but there are others things that i
wanta to put my energy like: server in node, deploy in cloud and
react concepts + tests. So i think tailwind could be a good match
to know this tecnology. I also put react-router-dom to start
spread the proeject in pages.